Adds logging for item dropped
