+++ Greedy.Spa/Pages/Game.razor
@@ -29,5 +29,4 @@
 
 </MudPaper>
 
-@* <DraggableDice DiceValues="@(DiceValues)"></DraggableDice> *@
-<DragabbleDice DiceValues="DiceValues"></DragabbleDice>
+<DragabbleDice DiceValues="DiceValues" OnDropCallback="@DieDropped"></DragabbleDice>
