+++ Greedy.WebApi/Application/GameService.cs
@@ -1,6 +1,5 @@
 using Eventuous;
 using Greedy.GameAggregate;
-using static Eventuous.CommandServiceDelegates;
 using static Eventuous.ExpectedState;
 
 namespace Greedy.WebApi.Application;
@@ -21,29 +20,5 @@ public class GameService : CommandService<Game, GameState, GameId> {
 
     On<Command.PassTurn>().InState(Existing).GetId(cmd => new GameId(cmd.GameId)).
       Execute((game, cmd) => game.PassTurn(cmd));
-  }
-}
-
-public static class CommandHandlerBuilderExtensions {
-  public static CommandHandlerBuilder<TCommand, TAggregate, TState, TId> Execute<TCommand, TAggregate, TState, TId>(
-    this CommandHandlerBuilder<TCommand, TAggregate, TState, TId> builder, ActOnAggregate<TAggregate, TCommand> action)
-    where TCommand : class
-    where TAggregate : Aggregate<TState>, new()
-    where TState : State<TState>, new()
-    where TId : Id
-  {
-    builder.Act((game, cmd) =>
-    {
-      try
-      {
-        action.Invoke(game, cmd);
-      }
-      catch (DomainException e)
-      {
-        // We ignore the domain exceptions because other wise the error events would not be persisted to the store. In a future version these events will be handled and will return the appropriate HTTP 400 Bad Request response
-      }
-    });
-
-    return builder;
   }
 }
